1-IngestionCapteurs/requirements.txt
python-dotenv
kafka-python
psycopg2-binary
